+++ src/pam.rs
@@ -3,141 +3,10 @@
 #![allow(non_camel_case_types)]
 
 use pam_types::PamHandle;
-use std::ffi::{CStr, CString};
 use std::fmt;
 
-#[cfg(feature = "libpam")]
-use libpam::{get_authtok, get_item, get_user, set_item};
-#[cfg(feature = "libpam")]
-use pam_types::{PamConv, PamItemType, PamMessage, PamMsgStyle, PamResponse};
-#[cfg(feature = "libpam")]
-use std::os::raw::{c_char, c_void};
-#[cfg(feature = "libpam")]
-use std::ptr;
-
-/// This contains a private marker trait, used to seal private traits.
-mod private {
-    pub trait Sealed {}
-    impl Sealed for super::Pam {}
-}
-
 /// Opaque PAM handle, with additional native methods available via `PamLibExt`.
-pub struct Pam(PamHandle);
-
-/// Extension trait over `Pam`, usually provided by the `libpam` shared library.
-pub trait PamLibExt: private::Sealed {
-    /// Get the username. If the PAM_USER item is not set, this function
-    /// prompts for a username (like get_authtok).
-    fn get_user(&self, prompt: Option<&str>) -> PamResult<Option<&CStr>>;
-
-    /// Get the username, i.e. the PAM_USER item. If it's not set return None.
-    fn get_cached_user(&self) -> PamResult<Option<&CStr>>;
-
-    /// Get the cached authentication token.
-    fn get_cached_authtok(&self) -> PamResult<Option<&CStr>>;
-
-    /// Get the cached authentication token or prompt the user for one if there isn't any.
-    fn get_authtok(&self, prompt: Option<&str>) -> PamResult<Option<&CStr>>;
-
-    fn set_authtok(&self, authtok: &CString) -> PamResult<()>;
-
-    /// Get the remote hostname.
-    fn get_rhost(&self) -> PamResult<Option<&CStr>>;
-
-    /// Get the remote username.
-    fn get_ruser(&self) -> PamResult<Option<&CStr>>;
-
-    #[cfg(feature = "libpam")]
-    /// Prompt the user for custom input.
-    fn conv(&self, prompt: Option<&str>, style: PamMsgStyle) -> PamResult<Option<&CStr>>;
-}
-
-#[cfg(feature = "libpam")]
-const ERR_CSTR_NULL: &str = "Error, the prompt cannot contain any null bytes";
-
-#[cfg(feature = "libpam")]
-impl PamLibExt for Pam {
-    fn get_user(&self, prompt: Option<&str>) -> PamResult<Option<&CStr>> {
-        let cprompt = prompt.map(|p| CString::new(p).expect(ERR_CSTR_NULL));
-        let pointer = get_user(self.0, cprompt.as_ref().map(|p| p.as_ptr()))?;
-        unsafe { Ok(pointer.map(|p| CStr::from_ptr(p))) }
-    }
-
-    fn get_cached_user(&self) -> PamResult<Option<&CStr>> {
-        let pointer = get_item(self.0, PamItemType::USER)?;
-        unsafe { Ok(pointer.map(|p| CStr::from_ptr(p as *const c_char))) }
-    }
-
-    fn get_cached_authtok(&self) -> PamResult<Option<&CStr>> {
-        // pam should keep the underlying token allocated for as long as the module is loaded
-        // which make this safe
-        unsafe {
-            let pointer = get_item(self.0, PamItemType::AUTHTOK)?;
-            Ok(pointer.map(|p| CStr::from_ptr(p as *const c_char)))
-        }
-    }
-
-    fn get_authtok(&self, prompt: Option<&str>) -> PamResult<Option<&CStr>> {
-        let cprompt = prompt.map(|p| CString::new(p).expect(ERR_CSTR_NULL));
-        let result = get_authtok(
-            self.0,
-            PamItemType::AUTHTOK,
-            cprompt.as_ref().map(|p| p.as_ptr()),
-        )?;
-        // If result is Ok we're guaranteed that p is a valid pointer
-        unsafe { Ok(result.map(|p| CStr::from_ptr(p))) }
-    }
-
-    fn set_authtok(&self, authtok: &CString) -> PamResult<()> {
-        unsafe {
-            set_item(
-                self.0,
-                PamItemType::AUTHTOK,
-                authtok.as_ptr() as *const c_void,
-            )
-        }
-    }
-
-    fn get_rhost(&self) -> PamResult<Option<&CStr>> {
-        let pointer = get_item(self.0, PamItemType::RHOST)?;
-        unsafe { Ok(pointer.map(|p| CStr::from_ptr(p as *const c_char))) }
-    }
-
-    fn get_ruser(&self) -> PamResult<Option<&CStr>> {
-        let pointer = get_item(self.0, PamItemType::RUSER)?;
-        unsafe { Ok(pointer.map(|p| CStr::from_ptr(p as *const c_char))) }
-    }
-
-    fn conv(&self, prompt: Option<&str>, style: PamMsgStyle) -> PamResult<Option<&CStr>> {
-        let pointer = match get_item(self.0, PamItemType::CONV)? {
-            Some(p) => p,
-            None => return Ok(None),
-        };
-        let conv = unsafe { &*(pointer as *const PamConv) };
-
-        let mut resp_ptr: *mut PamResponse = ptr::null_mut();
-        let msg_cstr = CString::new(prompt.unwrap_or("")).expect(ERR_CSTR_NULL);
-        let msg = PamMessage {
-            msg_style: style,
-            msg: msg_cstr.as_ptr(),
-        };
-
-        match conv.cb.map(|cb| {
-            PamError::new(cb(
-                1,
-                &mut (&msg as *const PamMessage),
-                &mut resp_ptr,
-                conv.appdata_ptr,
-            ))
-        }) {
-            Some(PamError::SUCCESS) => {
-                Ok(unsafe { (*resp_ptr).resp }.map(|r| unsafe { CStr::from_ptr(r.as_ptr()) }))
-            }
-            Some(ret) => Err(ret),
-            None => Ok(None),
-        }
-    }
-}
+pub struct Pam(pub(crate) PamHandle);
 
 pub type PamResult<T> = Result<T, PamError>;
